Restructed SD to have a header file to test
